% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_importance.R
\name{calculate_metric_score}
\alias{calculate_metric_score}
\title{Calculate metric score for a model and dataset}
\usage{
calculate_metric_score(model, h2o_data, target_column, metric)
}
\arguments{
\item{model}{H2O model object}

\item{h2o_data}{H2O frame}

\item{target_column}{Character. Name of target column}

\item{metric}{Character. Metric to calculate}
}
\value{
Numeric. The metric score
}
\description{
Calculate metric score for a model and dataset
}
\keyword{internal}
