% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_generation.R
\name{create_simple_html}
\alias{create_simple_html}
\title{Create simple HTML content}
\usage{
create_simple_html(imp_plot, model_plot, autoxplain_result, importance)
}
\arguments{
\item{imp_plot}{Plotly importance plot}

\item{model_plot}{Plotly model comparison plot}

\item{autoxplain_result}{AutoML results}

\item{importance}{Feature importance data}
}
\value{
Character vector with HTML content
}
\description{
Create simple HTML content
}
\keyword{internal}
