% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{get_feature_grid}
\alias{get_feature_grid}
\title{Generate feature grid for partial dependence calculation}
\usage{
get_feature_grid(feature_data, n_points, quantile_range)
}
\arguments{
\item{feature_data}{Numeric vector of feature values}

\item{n_points}{Integer. Number of points to generate}

\item{quantile_range}{Numeric vector of length 2. Quantile range}
}
\value{
Numeric vector of feature values to evaluate
}
\description{
Generate feature grid for partial dependence calculation
}
\keyword{internal}
