% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to check}

\item{y}{Default value to use if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Returns the first argument if it is not NULL, otherwise returns the second argument.
This is useful for providing default values.
}
\examples{
\dontrun{
# Returns "default"
NULL \%||\% "default"

# Returns "value"
"value" \%||\% "default"
}
}
\keyword{internal}
