% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_partial_dependence_multi}
\alias{plot_partial_dependence_multi}
\title{Plot multiple feature effects}
\usage{
plot_partial_dependence_multi(
  pdp_list,
  title = "Feature Effects",
  colors = NULL,
  ncol = 2L
)
}
\arguments{
\item{pdp_list}{Named list of feature effects.}

\item{title}{Overall title.}

\item{colors}{Line colors.}

\item{ncol}{Number of subplot columns.}
}
\value{
A Plotly htmlwidget.
}
\description{
Plot multiple feature effects
}
