% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_generation.R
\name{prepare_dashboard_data}
\alias{prepare_dashboard_data}
\title{Prepare data for dashboard generation}
\usage{
prepare_dashboard_data(autoxplain_result, top_features, sample_instances)
}
\arguments{
\item{autoxplain_result}{AutoML results}

\item{top_features}{Number of top features}

\item{sample_instances}{Number of sample instances}
}
\value{
List with prepared data
}
\description{
Prepare data for dashboard generation
}
\keyword{internal}
