% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_characterization.R
\name{print.autoxplainr_model_characteristics}
\alias{print.autoxplainr_model_characteristics}
\title{Print Method for Model Characteristics}
\usage{
\method{print}{autoxplainr_model_characteristics}(x, ...)
}
\arguments{
\item{x}{Object of class autoxplainr_model_characteristics}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for Model Characteristics
}
