% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_characterization.R
\name{summary.autoxplainr_model_characteristics}
\alias{summary.autoxplainr_model_characteristics}
\title{Summary Method for Model Characteristics}
\usage{
\method{summary}{autoxplainr_model_characteristics}(object, ...)
}
\arguments{
\item{object}{Object of class autoxplainr_model_characteristics}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for Model Characteristics
}
