% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_generation.R
\name{create_dashboard_rmd}
\alias{create_dashboard_rmd}
\title{Create flexdashboard Rmd content}
\usage{
create_dashboard_rmd(data_files)
}
\arguments{
\item{data_files}{List of file paths to data RDS files}
}
\value{
Character vector with Rmd content
}
\description{
Create flexdashboard Rmd content
}
\keyword{internal}
