% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_model_correlations}
\alias{plot_model_correlations}
\title{Plot prediction agreement among retained AutoML models}
\usage{
plot_model_correlations(autoxplain_result, test_data = NULL)
}
\arguments{
\item{autoxplain_result}{An \code{autoxplain_result}.}

\item{test_data}{Optional evaluation data.}
}
\value{
A Plotly heatmap.
}
\description{
Plot prediction agreement among retained AutoML models
}
